app-requirements.txt

tensorflow
fastapi
uvicorn
pillow
scikit-learn
python-multipart
psycopg2-binary
numpy
pandas
dotenv
streamlit
requests
pandas
plotly
python-dotenv
Pillow
locust
requests

